First Android project commit
This is the first commit that contains actual code. The source files
still need to be commented to make sense of the code, but eclipse should
be able to build and run this app. Please note that this contains
precompiled native code, so the native source files can't be edited.